+++ cpp/CHANGELOG.md
@@ -57,6 +57,7 @@ Please mark all change in change log and use the ticket from JIRA.
 - MS-415 - Add command tasktable to dump all tasktables
 - MS-418 - Update server_config.template file, set CPU compute only default
 - MS-419 - Move index_file_size from IndexParam to TableSchema
+- MS-421 - Add TaskLabel in scheduler
 
 ## New Feature
 - MS-343 - Implement ResourceMgr